ReadMe
In this assignment you will write a C++ program that reads a vocabulary, i.e. a set of words, from the std. input and (1) counts the number of occurrences of each word in the vocabulary, (2) prints out the words in the vocabulary in ascending order. The vocabulary size will not exceed 1000.
For example, if the vocabulary includes:
>Mehmet ayse MeHmet Ayse ayse MehMET barni mehmet
Your program will print to the std. output the following text:
Vocabulary Size = 3
ayse 3
barni 1
mehmet 4
Notice that, all words in the vocabulary should be treated case insensitively; and the words must be printed in small letters.